init: hello world, ensemble python!
Ensemble python is a library to orchestrate jobs based on
an algorithm in Python. The definitions are akin to state
machines, where starting jobs are defined, and then
rules with triggers and actions to take. The executor
queues are expected (and hoped) to be event driven so we
are not polling for anything. Flux does this nicely :)

Signed-off-by: vsoch <vsoch@users.noreply.github.com>