feat: support for repetitions and logging
Problem: we need to be able to define repetitions
and optionally backoff. I also wanted to add the ability
to turn off robust logging (that shows all events) and
better pretty print the output, and I did that here.
I also added support for flux job events, and updated
the README docs to point to the rfc for that. Likely
next I will work on the custom function, which I have
prototyped but need to implement. Then we can try
an actual dummy example with lammps and likely
integration into the flux operator!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>