feat: add support for terminate
Problem: in the context of the flux operator (or similar) we need
to be able to say "terminate on this condition." This addition
adds that. Since we are running a callback in the flux reactor,
we actually do not want to sys exit, but rather just stop the
reactor. It exits cleanly after that.
Solution: add a terminate action and queue function to call it.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>