feat: support for flux minicluster
Flux running in the MiniCluster contexts adds the cool
features to grow/shrink (autoscaling), which can be
paired with an actual cluster autoscaler (or not) if
the mini cluster is not using all the nodes available.
To support this we need a few things - first a heartbeat,
and one that runs at a user specified increment (and likely
this should also be exposed in the ensemble config) because
it is very likely the case that the triggers are not linked
to jobs (for example, "run this when the queue wait time is
over X for this job group"). We then need the MiniCluster
member, which is exactly the same as flux, but instead has
added the grow/shrink. I think likely what I want to do is
have the heartbeat disabled unless the setting is found in
the ensemble config OR the action to grow/shrink is found
(and we would use a default heartbeat seconds of 60). This
time should obviously be tested for different applications.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>