example: heartbeat example
This updates the heartbeat so it is entirely derived from
the config. This can happen explicitly if the user sets
logging->heartbeat to a non zero value, but it will also
happen if there is a grow or shrink action used. If the user
defines a grow/shrink and sets the heartbeat to 0 it will
still be set to the default, 60, because grow/shrink will
not work as expected without it.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>